refactor: chain -> shard
