+++ cardinal/ecs/options.go
@@ -1,10 +1,10 @@
 package ecs
 
-import "github.com/argus-labs/world-engine/cardinal/chain"
+import "github.com/argus-labs/world-engine/cardinal/shard"
 
 type Option func(w *World)
 
-func WithAdapter(adapter chain.Adapter) Option {
+func WithAdapter(adapter shard.Adapter) Option {
 	return func(w *World) {
 		w.chain = adapter
 	}
